+++ resources/views/components/app-fullwidth-layout.blade.php
@@ -1,0 +1,3 @@
+<div>
+    <!-- Order your soul. Reduce your wants. - Augustine -->
+</div>
